+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [yuyo.components.builder][] descriptor for adding a raw component builder to
+  a column without a callback.
+
 ### Changed
 - [yuyo.pagination.Page][] is now exported top-level at `Yuyo.Page`.